requirements.txt
# CDK and infrastructure
aws-cdk-lib>=2.150.0,<3
constructs>=10.0.0,<11

# Lambda + bridge runtime (for local testing)
boto3>=1.35.0
botocore>=1.35.0
litellm>=1.40.0
